Rewrite colour space classes

+++ src/maths/Maths.hpp
@@ -7,26 +7,26 @@ public:
   Maths() {};
   ~Maths() {};
 
-  static long double NRoot(const double v, const double n) {
-    long double out = v;
-    const long double exp = 1. / n;
+  static double NRoot(const double v, const double n) {
+    double out = v;
+    const double exp = 1. / n;
     if (std::fmod(n, 1.) == 0.) {
       out = std::pow(out, exp);
     }
     else {
-      long double absroot = std::pow(std::abs(out), exp);
+      double absroot = std::pow(std::abs(out), exp);
       if (out < 0.) absroot *= -1.;
       out = absroot;
     }
     return out;
   }
 
-  static long double UnsignedMod(const long double numer, const long double denom) {
+  static double UnsignedMod(const double numer, const double denom) {
     return std::fmod(std::fmod(numer, denom) + denom, denom);
   }
 
-  static const long double Pi;
-  static const long double Tau;
-  static const long double RadToDeg;
-  static const long double DegToRad;
+  static const double Pi;
+  static const double Tau;
+  static const double RadToDeg;
+  static const double DegToRad;
 };
